feat: changed folder structure for core [B1M2-3866]
