+++ packages/core/src/ageLimit/AgeLimit.tsx
@@ -1,28 +1,9 @@
 import * as React from 'react';
-import styled from 'styled-components';
-import { rem, themed } from '@heathmont/moon-utils';
+import { themed } from '@heathmont/moon-utils';
 import { ColorProps, useTheme } from '@heathmont/moon-themes';
 
-const Wrapper = styled.div(({ theme }) => ({
-  boxSizing: 'border-box',
-  border: `2px ${theme.color.chiChi[100]} solid`,
-  borderRadius: rem(theme.radius.largest),
-  padding: rem(theme.space.small),
-  width: rem(40),
-  height: rem(40),
-  display: 'flex',
-  justifyContent: 'center',
-  alignItems: 'center',
-  overflow: 'hidden',
-}));
-
-const Caption = styled.span(({ theme, color }) => ({
-  color,
-  fontSize: rem(14),
-  lineHeight: 1,
-  fontWeight: theme.fontWeight.semibold,
-}));
-
+import Container from './styles/Container';
+import Caption from './styles/Caption';
 interface AgeLimitProps {
   caption: any;
   color?: ColorProps;
@@ -35,9 +16,9 @@ const AgeLimit: React.FC<AgeLimitProps> = ({
   const theme = useTheme();
   const captionColor = themed('color', color)(theme);
   return (
-    <Wrapper>
+    <Container>
       <Caption color={captionColor}>{caption}</Caption>
-    </Wrapper>
+    </Container>
   );
 };
