--- packages/core/src/badge/README.mdx
@@ -1,126 +1,0 @@
----
-name: Badge
-menu: Design System
-route: /core/badge
----
-
-# Badge
-
-Small count and labeling component.
-
-## Usage:
-
-```
-import { Badge } from '@heathmont/moon-core';
-```
-
-```jsx react-live
-<Badge>Active</Badge>
-```
-
-## Props
-
-### Customize colours
-
-You could use `color` and `backgroundColor` props.
-
-Default colours are:
-
-```
-color="goten.100"
-
-backgroundColor="piccolo.100"
-```
-
-```jsx react-live
-<Inline>
-  <Badge color="goten.100" backgroundColor="piccolo.100">
-    Active
-  </Badge>
-  <Badge color="krillin.100" backgroundColor="gohan.100">
-    Active
-  </Badge>
-  <Badge color="goten.100" backgroundColor="dodoria.100">
-    Active
-  </Badge>
-</Inline>
-```
-
-### Customize size
-
-You could use `size="small"` prop.
-
-```
-size= "small" | undefined
-
-```
-
-```jsx react-live
-<Inline>
-  <Badge size="small">Small</Badge>
-  <Badge>Default</Badge>
-</Inline>
-```
-
-### Insert Icons to left, right or both sides
-
-```
-iconLeft={<IconProfile />}
-iconRight={<IconProfile />}
-```
-
-```jsx react-live
-<>
-  <Inline>
-    <Badge
-      size="small"
-      color="goten.100"
-      backgroundColor="piccolo.100"
-      iconLeft={<IconProfile />}
-    >
-      Active
-    </Badge>
-    <Badge
-      size="small"
-      color="goten.100"
-      backgroundColor="piccolo.100"
-      iconRight={<IconProfile />}
-    >
-      Active
-    </Badge>
-    <Badge
-      size="small"
-      color="goten.100"
-      backgroundColor="piccolo.100"
-      iconLeft={<IconProfile />}
-      iconRight={<IconProfile />}
-    >
-      Active
-    </Badge>
-  </Inline>
-  <Inline>
-    <Badge
-      color="goten.100"
-      backgroundColor="piccolo.100"
-      iconLeft={<IconProfile />}
-    >
-      Active
-    </Badge>
-    <Badge
-      color="goten.100"
-      backgroundColor="piccolo.100"
-      iconRight={<IconProfile />}
-    >
-      Active
-    </Badge>
-    <Badge
-      color="goten.100"
-      backgroundColor="piccolo.100"
-      iconLeft={<IconProfile />}
-      iconRight={<IconProfile />}
-    >
-      Active
-    </Badge>
-  </Inline>
-</>
-```
