--- packages/core/src/caption/README.mdx
@@ -1,53 +1,0 @@
----
-name: Caption
-menu: Design System
-route: /core/caption
----
-
-# Caption
-
-Caption is a Span tag with predefined styles. No option available except of color
-
-## color
-
-By default Caption color is bulma.100. You can use any color.
-
-Bulma.100 - is main text color. Trunks.100 - is secondary text color
-
-```jsx react-live
-<Stack>
-  <Caption>Caption with default color</Caption>
-  <Caption color="trunks.100">Caption with defined color</Caption>
-  <Caption color="piccolo.100">Caption with defined color</Caption>
-  <Caption color="krillin.100">Caption with defined color</Caption>
-</Stack>
-```
-
-## as="p"
-
-By default Caption renders as Span tag. You can render it as Paragraph or Label tags, etc.
-
-```jsx react-live
-<Stack>
-  <Caption>Caption renders Span tag by default</Caption>
-  <Caption as="p">Caption renders Paragraph tag</Caption>
-  <Caption as="label">Caption renders Label tag</Caption>
-</Stack>
-```
-
-## textAlign
-
-```jsx react-live
-<Stack>
-  <Caption textAlign="right">Caption renders Span tag by default</Caption>
-  <Caption dir="rtl" textAlign="right" as="p">
-    آراء وتحليلات.
-  </Caption>
-  <Caption dir="rtl" textAlign="right" color="piccolo.100">
-    Caption with defined color
-  </Caption>
-  <Caption dir="rtl" textAlign="right" color="krillin.100">
-    آراء وتحليلات.
-  </Caption>
-</Stack>
-```
