+++ packages/core/src/checkbox/@types/Variants.d.ts
@@ -1,0 +1,1 @@
+type Variants = 'option';
