--- packages/core/src/checkbox/README.mdx
@@ -1,120 +1,0 @@
----
-name: Checkbox
-menu: Design System
-route: /core/checkbox
----
-
-# Checkbox
-
-Checkboxes offer users a 'yes' or 'no' choice. They can be used on their own,
-or as a multiple choice fieldset.
-
-## Usage:
-
-```
-import { Checkbox } from '@heathmont/moon-core';
-```
-
-```jsx react-live
-<Checkbox label="I agree to receive bonus & marketing emails." />
-```
-
-## States
-
-### Disabled
-
-```jsx react-live
-<Checkbox disabled label="I agree to receive bonus & marketing emails." />
-```
-
-### Checked
-
-```jsx react-live
-<Checkbox checked label="I agree to receive bonus & marketing emails." />
-```
-
-## Props
-
-### Variants
-
-```jsx react-live
-<Checkbox
-  variant="option"
-  label="I agree to receive bonus & marketing emails."
-/>
-```
-
-### TODO
-
-### aria-label
-
-For use-cases where a label should be visually hidden, a descriptive `ariaLabel`
-must be defined. It's an attribute designed to help assistive technology (e.g. screen readers).
-
-```jsx react-live
-<Checkbox ariaLabel="Description of the checkbox's purpose." />
-```
-
-### ID
-
-`Checkbox` id can be overriden via `id` prop. That attribute helps old assistive technologies. (e.g. screen readers).
-
-```jsx react-live
-<Checkbox
-  id="example-id"
-  label="I agree to receive bonus & marketing emails."
-/>
-```
-
-# Composed Custom Checkbox API
-
-With custom checkbox, you can choose between a high-level API where DOM elements are not individually exposed as components,
-or use the composed API to access each sub-component directly.
-
-⚠️ Please note: Always use `CheckboxHiddenInput` before `CheckboxStyledInput` (due to internal css selectors order)
-
-```
-import {
-  CheckboxLabel,
-  CheckboxHiddenInput,
-  CheckboxStyledInput,
-  CheckboxCaption
-} from '@heathmont/moon-core';
-```
-
-```jsx react-live
-() => {
-  const disabled = false;
-  const id = 'test-checkbox-id';
-  const label = 'I agree to receive bonus & marketing emails.';
-  const ariaLabel = 'marketing emails';
-
-  return (
-    <CheckboxLabel htmlFor={id}>
-      <CheckboxHiddenInput disabled={disabled} aria-label={ariaLabel} id={id} />
-      <CheckboxStyledInput />
-      {label && <CheckboxCaption>{label}</CheckboxCaption>}
-    </CheckboxLabel>
-  );
-};
-```
-
-## Right aligned
-
-```jsx react-live
-() => {
-  const id2 = 'test-checkbox-right-id';
-  const label = 'آراء وتحليلات.';
-  const ariaLabel = 'marketing emails';
-
-  return (
-    <Inline justifyContent="flex-end">
-      <CheckboxLabel htmlFor={id2} dir="rtl">
-        {label && <CheckboxCaption>{label}</CheckboxCaption>}
-        <CheckboxHiddenInput aria-label={ariaLabel} id={id2} />
-        <CheckboxStyledInput />
-      </CheckboxLabel>
-    </Inline>
-  );
-};
-```
