+++ packages/core/src/checkbox/styles/CheckboxCaption.ts
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+const CheckboxCaption = styled.span(({ theme }) => ({
+  color: theme.color.trunks[100],
+}));
+
+export default CheckboxCaption;
