--- packages/core/src/heading/README.mdx
@@ -1,96 +1,0 @@
----
-name: Heading
-menu: Design System
-route: /core/heading
----
-
-# Heading
-
-## sizes
-
-We have predefined font sizes: 16 | 18 | 20 | 24 | 32 | 48 | 56 | 64 | 72. Line heights are calculated automatically based on font size.
-
-By default font size is 16. You can choose any size.
-
-```jsx react-live
-<Stack>
-  <Heading>Heading with default font size</Heading>
-  <Heading size={18}>Text with font size 18</Heading>
-  <Heading size={20}>Text with font size 20</Heading>
-  <Heading size={24}>Text with font size 24</Heading>
-  <Heading size={32}>Text with font size 32</Heading>
-  <Heading size={48}>font size 48</Heading>
-  <Heading size={56}>font size 56</Heading>
-  <Heading size={64}>font size 64</Heading>
-  <Heading size={72}>font size 72</Heading>
-</Stack>
-```
-
-## isRegular
-
-By default font weight is bold. You can make it thiner.
-
-```jsx react-live
-<Stack>
-  <Heading size={24}>Heading with default font weight</Heading>
-  <Heading size={24} isRegular>
-    Heading with font-weight: regular
-  </Heading>
-</Stack>
-```
-
-## as="h2"
-
-By default Heading renders as H3 tag. You can render it as H1, H2, H4, H5 and H6 tags.
-
-```jsx react-live
-<Stack>
-  <Heading size={24}>Heading renders H3 tag by default</Heading>
-  <Heading size={24} as="h2">
-    Heading renders H2 tag
-  </Heading>
-  <Heading size={24} as="h6">
-    Heading renders H6 tag
-  </Heading>
-</Stack>
-```
-
-## color
-
-By default Heading color is bulma.100. You can use any color.
-
-Bulma.100 - is main text color. Trunks.100 - is secondary text color
-
-```jsx react-live
-<Stack>
-  <Heading size={24}>Heading with default color</Heading>
-  <Heading size={24} color="trunks.100">
-    Heading with defined color
-  </Heading>
-  <Heading size={24} color="piccolo.100">
-    Heading with defined color
-  </Heading>
-  <Heading size={24} color="krillin.100">
-    Heading with defined color
-  </Heading>
-</Stack>
-```
-
-## textAlign
-
-```jsx react-live
-<Stack>
-  <Heading size={24} dir="rtl" textAlign="right">
-    آراء وتحليلات.
-  </Heading>
-  <Heading size={24} dir="rtl" textAlign="right" color="trunks.100">
-    Heading with defined color
-  </Heading>
-  <Heading size={24} textAlign="right" color="piccolo.100">
-    آراء وتحليلات.
-  </Heading>
-  <Heading size={24} textAlign="right" color="krillin.100">
-    آراء وتحليلات.
-  </Heading>
-</Stack>
-```
