+++ packages/core/src/index.ts
@@ -24,19 +24,9 @@ export { default as Search } from './search/Search';
 export * from './search/Search';
 export { default as TabLink } from './tabLink/TabLink';
 export * from './tabLink/TabLink';
-export { default as TabLinkContainer } from './tabLink/styles/TabLinkContainer';
-export * from './tabLink/styles/TabLinkContainer';
 export { default as TabLinkFill } from './tabLinkFill/TabLinkFill';
 export * from './tabLinkFill/TabLinkFill';
-export { default as TabLinkFillContainer } from './tabLinkFill/styles/TabLinkFillContainer';
-export * from './tabLinkFill/styles/TabLinkFillContainer';
 export { default as Tabs } from './tabs/Tabs';
 export * from './tabs/Tabs';
-export { default as SkipLink } from './tabs/styles/SkipLink';
-export * from './tabs/styles/SkipLink';
-export { default as TabList } from './tabs/styles/TabList';
-export * from './tabs/styles/TabList';
-export { default as TabNav } from './tabs/styles/TabNav';
-export * from './tabs/styles/TabNav';
 export { default as Text } from './text/Text';
 export * from './text/Text';
