+++ packages/core/src/listItem/styles/SubtextInner.ts
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+const SubtextInner = styled.div({
+  whiteSpace: 'nowrap',
+  overflow: 'hidden',
+  textOverflow: 'ellipsis',
+});
+
+export default SubtextInner;
