--- packages/core/src/loader/README.mdx
@@ -1,42 +1,0 @@
----
-name: Loader
-menu: Design System
-route: /core/loader
----
-
-# Loader
-
-Pure CSS loading animation.
-
-```jsx react-live
-<Inline>
-  <Loader />
-  <Loader color="piccolo.100" />
-</Inline>
-```
-
-## Usage
-
-### Customization
-
-The `color` prop accepts [themed color token keys](/themes/tokens) (in dot
-notation) as well as regular string color values; rgba, hex etc.
-
-```jsx react-live
-<Inline>
-  <Loader color="dodoria.100" />
-  <Loader color="trunks.100" />
-  <Loader color="#663399" />
-</Inline>
-```
-
-### Buttons
-
-```jsx react-live
-<Inline>
-  <Button variant="primary">Default</Button>
-  <Button variant="primary" progress>
-    Default
-  </Button>
-</Inline>
-```
