--- packages/core/src/screenLoader/README.mdx
@@ -1,53 +1,0 @@
----
-name: Screen Loader
-menu: Design System
-route: /core/Screen-loader
----
-
-# Ball Loader
-
-Pure CSS loading animation.
-
-## Usage:
-
-```
-import { ScreenLoader } from '@heathmont/moon-core';
-```
-
-```jsx react-live
-<Inline>
-  <ScreenLoader />
-</Inline>
-```
-
-## Props
-
-### Customize colours
-
-You could use `firstColor`, `secondColor`, `thirdColor` and `fourthColor` props.
-
-Default colours are:
-
-```
-firstColor="chiChi.100"
-secondColor="piccolo.100"
-thirdColor="cell.100"
-fourthColor="frieza.100"
-```
-
-```jsx react-live
-<Inline>
-  <ScreenLoader
-    firstColor="piccolo.100"
-    secondColor="piccolo.100"
-    thirdColor="piccolo.100"
-    fourthColor="piccolo.100"
-  />
-  <ScreenLoader
-    firstColor="red"
-    secondColor="piccolo.100"
-    thirdColor="piccolo.100"
-    fourthColor="red"
-  />
-</Inline>
-```
