+++ packages/core/src/tabLinkFill/TabLinkFill.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from 'styled-components';
 import { rem } from '@heathmont/moon-utils';
 
-import TabLinkFillContainer from './styles/TabLinkFillContainer';
+import Container from './styles/Container';
 
 export type TabLinkFillProps = {
   elementLeft?: React.ReactElement;
@@ -27,11 +27,11 @@ const TabLinkFill: React.FC<TabLinkFillProps> = ({
   size,
   count,
 }) => (
-  <TabLinkFillContainer size={size} href={href}>
+  <Container size={size} href={href}>
     {elementLeft && <ElementLeftWrapper>{elementLeft}</ElementLeftWrapper>}
     {children}
     {count && <CountWrapper>{count}</CountWrapper>}
-  </TabLinkFillContainer>
+  </Container>
 );
 
 export default TabLinkFill;
