--- packages/core/src/text/README.mdx
@@ -1,92 +1,0 @@
----
-name: Text
-menu: Design System
-route: /core/text
----
-
-# Text
-
-We treat semantic and visual/cosmetic decisions as two totally [separate](https://csswizardry.com/2016/02/managing-typography-on-large-apps/) things.
-
-Our choice of `p` and `span` is a purely semantic decision, and doesn't impact cosmetics configured via the `size` prop.
-
-## sizes
-
-We have predefined font sizes: 10 | 12 | 14 | 16 | 18 | 20 | 24 | 32 | 48 | 56 | 64 | 72. Line heights are calculated automatically based on font size.
-
-By default font size is 16. You can choose any size.
-
-```jsx react-live
-<Stack>
-  <Text>Text with default font size</Text>
-  <Text size={10}>Text with font size 10</Text>
-  <Text size={12}>Text with font size 12</Text>
-  <Text size={14}>Text with font size 14</Text>
-  <Text size={16}>Text with font size 16</Text>
-  <Text size={18}>Text with font size 18</Text>
-  <Text size={20}>Text with font size 20</Text>
-  <Text size={24}>Text with font size 24</Text>
-  <Text size={32}>Text with font size 32</Text>
-  <Text size={48}>font size 48</Text>
-  <Text size={56}>font size 56</Text>
-  <Text size={64}>font size 64</Text>
-  <Text size={72}>font size 72</Text>
-</Stack>
-```
-
-## isBold
-
-By default font weight is regular. You can make it bolder.
-
-```jsx react-live
-<Stack>
-  <Text>Text with default font weight</Text>
-  <Text isBold>Text with font-weight: semibold</Text>
-</Stack>
-```
-
-## as="span"
-
-By default Text renders as Paragraph tag. You can render it as Span or Label tags, etc.
-
-```jsx react-live
-<Stack>
-  <Text>Text renders Paragraph tag by default</Text>
-  <Text as="span">Text renders Span tag</Text>
-  <Text as="label">Text renders Label tag</Text>
-</Stack>
-```
-
-## color
-
-By default Text color is bulma.100. You can use any color.
-
-Bulma.100 - is main text color. Trunks.100 - is secondary text color
-
-```jsx react-live
-<Stack>
-  <Text>Text with default color</Text>
-  <Text color="trunks.100">Text with defined color</Text>
-  <Text color="piccolo.100">Text with defined color</Text>
-  <Text color="krillin.100">Text with defined color</Text>
-</Stack>
-```
-
-## textAlign
-
-```jsx react-live
-<Stack>
-  <Text textAlign="right" dir="rtl">
-    آراء وتحليلات.
-  </Text>
-  <Text textAlign="right" color="trunks.100">
-    Text with defined color
-  </Text>
-  <Text textAlign="right" dir="rtl" color="piccolo.100">
-    آراء وتحليلات.
-  </Text>
-  <Text textAlign="right" dir="rtl" color="krillin.100">
-    آراء وتحليلات.
-  </Text>
-</Stack>
-```
